backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "groq>=1.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
